requirements.txt
Django==5.0.4
nltk==3.7
numpy==1.26.4
pandas==1.5.3
Pillow==10.3.0
python-dotenv==1.0.1
Requests==2.31.0
scipy==1.14.0
sentence_transformers==2.6.1
transformers==4.39.3
huggingface-hub==0.22.2